Use db instead of ndb due to problems eliminating sessions.

+++ gaesessions/__init__.py
@@ -1,22 +1,26 @@
 """A fast, lightweight, and secure session WSGI middleware for use with GAE."""
 from Cookie import CookieError, SimpleCookie
 from base64 import b64decode, b64encode
-import binascii
+from google.appengine.api.datastore import entity_pb
 import datetime
+import binascii
 import hashlib
 import hmac
-import logging
 import pickle
 import os
 import time
 import webapp2
 
-from google.appengine.ext import ndb
+from google.appengine.api import memcache
+
+# Do not try to update to ndb - see a description of the bug that prevents using ndb
+# here: http://stackoverflow.com/questions/15437320/appengine-python-ndb-key-order
+from google.appengine.ext import db
 from google.appengine.api import taskqueue
-from google.appengine.api.datastore import entity_pb
 
 from video_src import status_reporting
 
+import logging
 # ARM Note: We do not allow cookie-only session so that if we need to eliminate a user from the system and immediately
 # revoke their website access, this can be achieved by just removing their session from the datastore.
 # This should be revisited in the future to see if there are other methods of achieving this functionality
@@ -24,7 +28,7 @@ from video_src import status_reporting
 # amount of data between the client and the server.
 
 # Configurable cookie options
-COOKIE_NAME_PREFIX = "lx-session-"  # identifies a cookie as being one used by gae-sessions (so you can set cookies too)
+COOKIE_NAME_PREFIX = "cs-"  # identifies a cookie as being one used by gae-sessions (so you can set cookies too)
 COOKIE_PATH = "/"
 # Original DEFAULT_COOKIE_ONLY_THRESH was 10240 10KB: GAE only allows ~16000B in HTTP header - leave ~6KB for other info
 # However, we set to 0 because we don't want cookie-only sessions - we need to be able to remotely kills sessions 
@@ -33,7 +37,7 @@ DEFAULT_COOKIE_ONLY_THRESH = 0
 DEFAULT_LIFETIME = datetime.timedelta(hours=24)
 
 # constants
-SID_LEN = 43  # timestamp (10 chars) + underscore + urandom (32 hex chars)
+SID_LEN = 43  # timestamp (10 chars) + underscore + random(32 hex chars)
 SIG_LEN = 44  # base 64 encoded HMAC-SHA256
 MAX_COOKIE_LEN = 4096
 EXPIRE_COOKIE_FMT = ' %s=; expires=Wed, 01-Jan-1970 00:00:00 GMT; Path=' + COOKIE_PATH
@@ -48,10 +52,10 @@ def is_gaesessions_key(k):
     return k.startswith(COOKIE_NAME_PREFIX)
 
 
-class SessionModel(ndb.Model):
-    """Contains session data.  id is the session ID and pdump contains a
+class SessionModel(db.Model):
+    """Contains session data.  key_name is the session ID and pdump contains a
     pickled dictionary which maps session variables to their values."""
-    pdump = ndb.BlobProperty()
+    pdump = db.BlobProperty()
 
 
 class Session(object):
@@ -62,7 +66,7 @@ class Session(object):
     """
     DIRTY_BUT_DONT_PERSIST_TO_DB = 0
 
-    def __init__(self, sid=None, lifetime=DEFAULT_LIFETIME,
+    def __init__(self, sid=None, lifetime=DEFAULT_LIFETIME, no_datastore=False,
                  cookie_only_threshold=DEFAULT_COOKIE_ONLY_THRESH, cookie_key=None):
         self._accessed = False
         self.sid = None
@@ -72,6 +76,7 @@ class Session(object):
         self.dirty = False  # has the session been changed?
 
         self.lifetime = lifetime
+        self.no_datastore = no_datastore
         self.cookie_only_thresh = cookie_only_threshold
         self.base_key = cookie_key
 
@@ -110,7 +115,7 @@ class Session(object):
                 if pdump:
                     self.data = self.__decode_data(pdump)
                 else:
-                    self.data = None  # data is in db: load it on-demand
+                    self.data = None  # data is in memcache/db: load it on-demand
             else:
                 logging.warn('cookie with invalid sig received from %s: %s' % (os.environ.get('REMOTE_ADDR'), b64pdump))
         except (CookieError, KeyError, IndexError, TypeError):
@@ -196,14 +201,14 @@ class Session(object):
 
     @staticmethod
     def __encode_data(d):
-        """Returns a "pickled+" encoding of d.  d values of type ndb.Model are
+        """Returns a "pickled+" encoding of d.  d values of type db.Model are
         protobuf encoded before pickling to minimize CPU usage & data size."""
-        # separate protobufs so we'll know how to decode (they are just strings)
+        # separate protobufs so we'll know how to decode
         eP = {}  # for models encoded as protobufs
         eO = {}  # for everything else
         for k, v in d.iteritems():
-            if isinstance(v, ndb.Model):
-                eP[k] = ndb.ModelAdapter().entity_to_pb(v).Encode()
+            if isinstance(v, db.Model):
+                eP[k] = db.model_to_protobuf(v).Encode()
             else:
                 eO[k] = v
         return pickle.dumps((eP, eO), 2)
@@ -214,12 +219,11 @@ class Session(object):
         try:
             eP, eO = pickle.loads(pdump)
             for k, v in eP.iteritems():
-                eO[k] = ndb.ModelAdapter().pb_to_entity(entity_pb.EntityProto(v))
+                eO[k] = db.model_from_protobuf(entity_pb.EntityProto(v))
         except Exception, e:
             logging.warn("failed to decode session data: %s" % e)
             eO = {}
         return eO
-
 
     def regenerate_id(self, expiration_ts=None):
         """Assigns the session a new session ID (data carries over).  This
@@ -273,37 +277,43 @@ class Session(object):
         if self.sid:
             self.__clear_data()
         self.sid = sid
-        self.ndb_key = ndb.Key(SessionModel._get_kind(), sid, namespace='')
+        self.db_key = db.Key.from_path(SessionModel.kind(), sid, namespace='')
 
         # set the cookie if requested
         if make_cookie:
             self.cookie_data = ''  # trigger the cookie to be sent
 
     def __clear_data(self):
-        """Deletes this session from the datastore."""
-        logging.info("erasing session from database  %s" % self.sid)
+        """Deletes this session from memcache and the datastore."""
+        logging.info("erasing session from database and memcache %s" % self.sid) 
         if self.sid:
+            memcache.delete(self.sid, namespace='')  # not really needed; it'll go away on its own
             try:
-                self.ndb_key.delete()
+                db.delete(self.db_key)
             except:
-                pass  # it wasn't in the db (maybe cookie only or db is down) => cron will expire it
+                pass  # either it wasn't in the db (maybe cookie/memcache-only) or db is down => cron will expire it
 
     def __retrieve_data(self):
         """Sets the data associated with this session after retrieving it from
-        the datastore.  Assumes self.sid is set.  Checks for session
+        memcache or the datastore.  Assumes self.sid is set.  Checks for session
         expiration after getting the data."""
-
-        session_model_instance = self.ndb_key.get()
-        if session_model_instance:
-            pdump = session_model_instance.pdump
-        else:
-            expiry_datetime = self.get_expiration_datetime()
-            # This can happen if the user has multiple windows open, and logs out in one of the windows but
-            # the other windows continue to request information.
-            logging.warning("can't find session data in the datastore for sid=%s key = %s expiry: %s" % (self.sid, self.ndb_key, expiry_datetime))
-            self.terminate(False)  # we lost it; just kill the session
-            return
-
+        pdump = memcache.get(self.sid, namespace='')
+        if pdump is None:
+            # memcache lost it, go to the datastore
+            if self.no_datastore:
+                logging.warning("can't find session data in memcache for sid=%s (using memcache only sessions)" % self.sid)
+                self.terminate(False)  # we lost it; just kill the session
+                return
+            session_model_instance = db.get(self.db_key)
+            if session_model_instance:
+                pdump = session_model_instance.pdump
+            else:
+                expiry_datetime = self.get_expiration_datetime()
+                # This can happen if the user has multiple windows open, and logs out in one of the windows but 
+                # the other windows continue to request information.   
+                logging.warning("can't find session data in the datastore for sid=%s key = %s expiry: %s" % (self.sid, self.db_key, expiry_datetime))
+                self.terminate(False)  # we lost it; just kill the session
+                return
         self.data = self.__decode_data(pdump)
 
     def save(self, persist_even_if_using_cookie=True):
@@ -312,8 +322,9 @@ class Session(object):
         is called).
 
         If the data is small enough it will be sent back to the user in a cookie
-        instead of using the datastore.  If `persist_even_if_using_cookie`
-        evaluates to True, the datastore will also be used.
+        instead of using memcache and the datastore.  If `persist_even_if_using_cookie`
+        evaluates to True, memcache and the datastore will also be used.  If the
+        no_datastore option is set, then the datastore will never be used.
 
         Normally this method does not need to be called directly - a session is
         automatically saved at the end of the request if any changes were made.
@@ -342,10 +353,14 @@ class Session(object):
             # latest data will only be in the backend, so expire data cookies we set            
             self.cookie_data = ''
 
+        memcache.set(self.sid, pdump, namespace='', time=self.get_expiration())  # may fail if memcache is down
 
+        #if dirty is Session.DIRTY_BUT_DONT_PERSIST_TO_DB or self.no_datastore:
+            #return
+            
         # persist the session to the datastore            
         try:
-            SessionModel(id=self.sid, pdump=pdump).put()
+            SessionModel(key_name=self.sid, pdump=pdump).put()
         except Exception, e:
             logging.error("unable to persist session to datastore for sid=%s (%s)" % (self.sid, e))
 
@@ -372,6 +387,23 @@ class Session(object):
         self.dirty = True
         return self.data.pop(key, default)
 
+    def pop_quick(self, key, default=None):
+        """Removes key and returns its value, or default if key is not present.
+        The change will only be persisted to memcache until another change
+        necessitates a write to the datastore."""
+        self.ensure_data_loaded()
+        if self.dirty is False:
+            self.dirty = Session.DIRTY_BUT_DONT_PERSIST_TO_DB
+        return self.data.pop(key, default)
+
+    def set_quick(self, key, value):
+        """Set a value named key on this session.  The change will only be
+        persisted to memcache until another change necessitates a write to the
+        datastore.  This will start a session if one is not already active."""
+        dirty = self.dirty
+        self[key] = value
+        if dirty is False or dirty is Session.DIRTY_BUT_DONT_PERSIST_TO_DB:
+            self.dirty = Session.DIRTY_BUT_DONT_PERSIST_TO_DB
 
     def __getitem__(self, key):
         """Returns the value associated with key on this session."""
@@ -411,7 +443,6 @@ class Session(object):
         else:
             return "uninitialized session"
 
-
 class SessionAdmin(webapp2.RequestHandler):
 
     def delete_expired_sessions(self):
@@ -419,12 +450,12 @@ class SessionAdmin(webapp2.RequestHandler):
         If there are more than 500 expired sessions, only 500 will be removed.
         """
         now_str = unicode(int(time.time()))
-        q = SessionModel.query()
-        k = ndb.Key('SessionModel', now_str + u'\ufffd')
-        q.filter(SessionModel._key < k)
-        results = q.fetch(500, keys_only=True,)
+        q = db.Query(SessionModel, keys_only=True, namespace='')
+        key = db.Key.from_path('SessionModel', now_str + u'\ufffd', namespace='')
+        q.filter('__key__ < ', key)
+        results = q.fetch(500)
         num_results = len(results)
-        ndb.delete_multi([k for k in results])
+        db.delete(results)
         return num_results
